Room cloning

+++ yatsim_dashboard/views.py
@@ -4,7 +4,7 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect, render
 from django.views.generic.edit import FormView
 
-from .forms import RoomCreationForm
+from .forms import RoomCloneForm, RoomCreationForm, RoomDeletionForm
 from .models import Room
 
 # Create your views here.
@@ -42,6 +42,61 @@ class CreateRoomView(FormView):
 
 @login_required
 def delete_room(request, room_id):
-    user = request.user
-    Room.objects.filter(id__exact=room_id).filter(owner__exact=user).delete()
+    if request.method == "POST":
+        user = request.user
+        form = RoomDeletionForm(request.POST, request.FILES)
+        if form.is_valid():
+            Room.objects.filter(id__exact=room_id).filter(owner__exact=user).delete()
+        else:
+            pass
     return redirect("/dashboard")
+
+
+@login_required
+def clone_room(request, room_id):
+    if request.method == "POST":
+        user = request.user
+        form = RoomCloneForm(request.POST, request.FILES)
+        if form.is_valid():
+            try:
+                room = Room.objects.get(id=room_id)
+                print(room.owner)
+                print(user)
+                if room.owner == user:
+                    new_room = Room.objects.create(
+                        owner=user,
+                        room_name=request.POST["room_name"],
+                        height=room.height,
+                        width=room.width,
+                    )
+                    new_room.save()
+            except:
+                pass
+        else:
+            pass
+    return redirect("/dashboard")
+
+
+@login_required
+def place_cell(request):
+    pass
+
+
+@login_required
+def switch_cell(request):
+    pass
+
+
+@login_required
+def rotate_cell(request):
+    pass
+
+
+@login_required
+def add_user_to_room(request):
+    pass
+
+
+@login_required
+def remove_user_to_room(request):
+    pass
